Make configuration DB location survive ESXi reboot
On reboot all changes in /etc are lost, unless they are backed up.
ESXi has a standard backup for /etc/rc.local.d/local.sh - this file is
backed up and run on boot ,so to persist symlink to shared DB we need to
insert the link crestion to local.sh. And when we drop the link, we need
to clean up the local.sh

This is exactly what this change is doing